# openapi.yaml
openapi: 3.0.3
info:
  title: NutriCare Connect API
  version: 1.0.0
  description: Local portfolio application API. Seed records are fictional, health thresholds are non-diagnostic, and payment gateways remain simulated. SMTP email can be enabled locally.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Access]
      security: []
      summary: Register a user
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } } }
      responses: { '201': { description: Account created }, '400': { $ref: '#/components/responses/BadRequest' } }
  /auth/login:
    post:
      tags: [Access]
      security: []
      summary: Authenticate and issue a JWT
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email,password], properties: { email: { type: string, format: email }, password: { type: string, format: password } } } } } }
      responses: { '200': { description: JWT and user view }, '400': { $ref: '#/components/responses/BadRequest' } }
...
  /auth/reset-password:
    post:
      tags: [Access]
      security: []
      summary: Reset a password using a valid ten-minute code
      responses: { '204': { description: Password changed }, '400': { $ref: '#/components/responses/BadRequest' } }
...
  /users/me:
    get: { tags: [Access], summary: Get the signed-in user's profile, responses: { '200': { description: Current profile } } }
    patch:
      tags: [Access]
      summary: Update changeable profile information
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileRequest' } } } }
      responses: { '200': { description: Updated profile } }
...
  /users/staff:
    post:
      tags: [Access]
      summary: Provision a staff account and return its one-time temporary password
      responses: { '201': { description: Staff user and one-time password }, '403': { description: System administrator role required } }
  /users/{id}/status:
    patch:
      tags: [Access]
      summary: Enable or disable an account without deleting linked records
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [enabled], properties: { enabled: { type: boolean } } } } } }
      responses: { '200': { description: Updated user } }
  /users/{id}/role:
    patch:
      tags: [Access]
      summary: Change an assigned role (system administrator only)
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      responses: { '200': { description: Updated user } }
  /users/{id}/unlock:
    post:
      tags: [Access]
      summary: Unlock an account
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      responses: { '200': { description: Updated user } }
  /schedules/slots:
    get:
      tags: [Appointments]
      summary: List slots for a date
      parameters: [ { name: date, in: query, required: true, schema: { type: string, format: date } } ]
      responses: { '200': { description: Availability slots } }
  /appointments/hold:
    post:
      tags: [Appointments]
      summary: Hold a slot for ten minutes and create an invoice
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/HoldRequest' } } } }
      responses: { '201': { description: Booking, slot and invoice }, '409': { description: Slot conflict } }
  /appointments/{id}/payments:
    post:
      tags: [Appointments]
      summary: Save a simulated payment result
      description: Card numbers and CVV fields are deliberately unsupported.
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentRequest' } } } }
      responses: { '201': { description: Simulated payment record } }
  /appointments/{id}/cancel:
    post:
      tags: [Appointments]
      summary: Cancel and release a booking slot
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses: { '200': { description: Cancelled } }
  /appointments/patient/{id}:
    get:
      tags: [Appointments]
      summary: List patient appointments
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      responses: { '200': { description: Appointments } }
...
  /checkups:
    post:
      tags: [Health]
      summary: Record a health check and create threshold alerts
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CheckRequest' } } } }
      responses: { '201': { description: Check, alerts and disclaimer } }
  /checkups/patient/{id}:
    get:
      tags: [Health]
      summary: Get patient health history
      parameters: [ { $ref: '#/components/parameters/AccountId' } ]
      responses: { '200': { description: Health checks } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    AccountId: { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/AccountId' } }
  responses:
    BadRequest: { description: Validation or state rejection }
  schemas:
    AccountId: { type: string, pattern: '^[A-Z]{1,2}[0-9]{3}$', example: P001 }
    RegisterRequest:
      type: object
      required: [fullName,email,password]
      properties: { fullName: { type: string }, email: { type: string, format: email }, password: { type: string, minLength: 8, maxLength: 72 }, phoneNumber: { type: string }, dateOfBirth: { type: string, format: date } }
    ProfileRequest:
      type: object
      required: [fullName,email]
      properties: { fullName: { type: string }, email: { type: string, format: email }, phoneNumber: { type: string }, dateOfBirth: { type: string, format: date }, address: { type: string, maxLength: 500 } }
    HoldRequest:
      type: object
      required: [slotId,patientId,serviceType,amount]
      properties: { slotId: { type: string, format: uuid }, patientId: { $ref: '#/components/schemas/AccountId' }, serviceType: { type: string }, amount: { type: number, format: decimal, minimum: 0 } }
    PaymentRequest:
      type: object
      required: [amount,method,status]
      properties:
        amount: { type: number, minimum: 0 }
        method: { type: string, enum: [CARD,CASH,BANK_TRANSFER,MOBILE_WALLET] }
        status: { type: string, enum: [PAID,PENDING,PARTIALLY_PAID,FAILED,REFUNDED] }
    CheckRequest:
      type: object
      required: [patientId]
      description: The practitioner ID is derived from the authenticated account and cannot be supplied by the client.
      properties: { patientId: { $ref: '#/components/schemas/AccountId' }, weightKg: { type: number }, bmi: { type: number }, systolic: { type: integer }, diastolic: { type: integer }, bloodSugar: { type: number }, temperature: { type: number }, notes: { type: string } }
...